Add HADOOP_HOME env var and adds bin to defalt path
This patch adds missing HADOOP_HOME variable which
is used by other software such as Apache mesos.
And adds the hadoop bin folder to the default path
of all users. This way every user in the
system is able to execute hadoop commands.

This patch also fixes a bug where the HADOOP_PREFIX
variable was hard coded where it should have been using
the settings of the user / default defined in
hadoop/settings.sls